refactor(core): Move Agent schemas out
